Struggling with mutually recursive parsers
Mutually recursive functions aren't a problem in purescript, but
mutually recursive parsers are. The data definitions reference one
another, which isn't allowed in the javascript runtime. I tried
wrapping the recursive parsers in delayed evaluation but that doesn't
actually work... it delays evaluation, but only until we try to
instantiate the parses, and then it tries to create them, producing
the same cycle that the compiler forbids without the delayed
evaluation and thereby blowing the stack.

I'm still trying to understand how to use the suggestions in
https://github.com/purescript/documentation/blob/master/errors/CycleInDeclaration.md